+++ pom.xml
@@ -48,6 +48,7 @@
         <junit.version>4.7</junit.version>
         <hibernate.version>5.1.0.Final</hibernate.version>
         <guice.version>4.0</guice.version>
+        <polyguice.version>1.0</polyguice.version>
         <javax.inject.version>1</javax.inject.version>
         <hystrix.version>1.5.1</hystrix.version>
         <akka.version>2.4.1</akka.version>
