+++ runtime/src/test/java/com/flipkart/flux/dao/DummyEventData.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2012-2016, the original author or authors.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.flipkart.flux.dao;
+
+import java.io.Serializable;
+
+/**
+ * Dummy class which would be used as {@link com.flipkart.flux.domain.Event}'s eventData.
+ * @author shyam.akirala
+ */
+public class DummyEventData implements Serializable {
+    String data;
+
+    public DummyEventData(String data) {
+        this.data = data;
+    }
+
+    public String getData() {
+        return data;
+    }
+
+    public void setData(String data) {
+        this.data = data;
+    }
+
+}
