+++ runtime/src/test/java/com/flipkart/flux/dao/StateMachinesDAOTest.java
@@ -26,6 +26,7 @@ import com.google.inject.Injector;
 import junit.framework.Assert;
 
 /**
+ * <code>StateMachinesDAOTest</code> class tests the functionality of {@link StateMachinesDAO} using JUnit tests.
  * @author shyam.akirala
  */
 public class StateMachinesDAOTest {
@@ -43,32 +44,15 @@ public class StateMachinesDAOTest {
         DummyTask task = new DummyTask();
         DummyOnEntryHook onEntryHook = new DummyOnEntryHook();
         DummyOnExitHook onExitHook = new DummyOnExitHook();
-        State<Data> state1 = new State<Data>(2L, "state1", "desc1", onEntryHook, task, onExitHook, 3L, 60L);
-        State<Data> state2 = new State<Data>(2L, "state2", "desc2", null, null, null, 2L, 50L);
-        Set<State<Data>> states = new HashSet<State<Data>>();
+        State<DummyEventData> state1 = new State<DummyEventData>(2L, "state1", "desc1", onEntryHook, task, onExitHook, 3L, 60L);
+        State<DummyEventData> state2 = new State<DummyEventData>(2L, "state2", "desc2", null, null, null, 2L, 50L);
+        Set<State<DummyEventData>> states = new HashSet<State<DummyEventData>>();
         states.add(state1);
         states.add(state2);
-        StateMachine<Data> stateMachine = new StateMachine<Data>(2L, "test_name", "test_desc", states);
+        StateMachine<DummyEventData> stateMachine = new StateMachine<DummyEventData>(2L, "test_name", "test_desc", states);
         String savedSMId = stateMachinesDAO.create(stateMachine).getId();
 
         StateMachine stateMachine1 = stateMachinesDAO.findById(savedSMId);
         Assert.assertNotNull(stateMachine1);
-    }
-}
-
-/** Event data class*/
-class Data {
-    String data;
-
-    public Data(String data) {
-        this.data = data;
-    }
-
-    public String getData() {
-        return data;
-    }
-
-    public void setData(String data) {
-        this.data = data;
     }
 }
